+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/audio/AudioTrackAudioOutputProvider.java
@@ -265,7 +265,7 @@ public final class AudioTrackAudioOutputProvider implements AudioOutputProvider 
       outputMode = OUTPUT_MODE_PCM;
       outputEncoding = format.pcmEncoding;
       outputSampleRate = format.sampleRate;
-      outputChannelConfig = getAudioOutputChannelConfig(format.channelCount);
+      outputChannelConfig = getAudioOutputChannelConfig(format);
       outputPcmFrameSize = Util.getPcmFrameSize(outputEncoding, format.channelCount);
       usePlaybackParameters = formatConfig.enablePlaybackParameters;
     } else {
@@ -279,7 +279,7 @@ public final class AudioTrackAudioOutputProvider implements AudioOutputProvider 
       if (formatConfig.enableOffload && audioOffloadSupport.isFormatSupported) {
         outputMode = OUTPUT_MODE_OFFLOAD;
         outputEncoding = MimeTypes.getEncoding(checkNotNull(format.sampleMimeType), format.codecs);
-        outputChannelConfig = getAudioOutputChannelConfig(format.channelCount);
+        outputChannelConfig = getAudioOutputChannelConfig(format);
         // Offload requires AudioTrack playback parameters to apply speed changes quickly.
         usePlaybackParameters = true;
         useOffloadGapless = audioOffloadSupport.isGaplessSupported;
@@ -465,12 +465,12 @@ public final class AudioTrackAudioOutputProvider implements AudioOutputProvider 
     }
   }
 
-  private int getAudioOutputChannelConfig(int channelCount) {
+  private int getAudioOutputChannelConfig(Format format) {
     if (audioTrackProvider != null) {
-      return audioTrackProvider.getAudioTrackChannelConfig(channelCount);
+      return audioTrackProvider.getAudioTrackChannelConfig(format.channelCount);
     }
 
-    return Util.getAudioTrackChannelConfig(channelCount);
+    return Util.getAudioTrackChannelConfig(format);
   }
 
   private int getAudioTrackMinBufferSize(int sampleRateInHz, int channelConfig, int encoding) {